split main circuits
