+++ circuits/circuits/tests/isValid_tester.circom
@@ -1,0 +1,17 @@
+pragma circom  2.1.6;
+
+include "../IsValid.circom";
+
+template IsValid_tester() {
+
+    signal input currDate[6];
+    signal input validityDateASCII[6];
+    
+    component isValid = IsValid();
+    isValid.currDate <== currDate;
+    isValid.validityDateASCII <== validityDateASCII;
+
+    signal output out <== isValid.out;
+}
+
+component main = IsValid_tester();
